# [PERF] account: speedup `_count_results_and_sum_amounts` in dashboard