## Description
After importing a large quantity of pending bills with no `invoice_date`, the dashboard can be significantly slow.
The bottleneck comes primarily from the hot loop in `_count_results_and_sum_amounts`, as it's iterating over a potentially enormous list of dictionary of query results.
This patch improvement consists mostly of hoisting invariants out of this hot loop, most impactful is the call to `context_today`, which does a lot of computation/conversions to apply the proper timezone and localize the date.

## Benchmark

| len(results_dict) | Before | After | Speedup |
|-------------------|--------|-------|---------|
| Timings (10k)     | 250ms  | 50ms  | 5x      |
| Timings (50k)     | 1.2s   | 300ms | 4x      |
| Timings (100k)    | 2.7s   | 820ms | 3.3x    |
| Timings (250k)    | 6.7s   | 1.7s  | 3.9x    |
| Timings (~600k)   | 13s    | 4s    | 3.25x   |

## Reference
opw-4235252

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr