@robodoo 2024-10-11
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/183342.png)](https://mergebot.odoo.com/odoo/odoo/pull/183342)