@pivi-odoo 2024-10-11
> Nice catch on the `context_today`. 👍 Since you basically change the whole function anyway have you tried moving the `document_currency` and `date` into the `else` branch? (not sure it makes a difference) Or maybe the diff could be reduced a bit? (I get that performant code is not necessarily "nice" and it is fine for me generally speaking)

Hmm, I've just tried your suggestion, as it seemed a good idea to move them to the else branch, but after re-benching, for the last case (the largest), and funnily enough, it's *consistently* slower (by around 400~500ms). As for the reason, I have no idea why, probably a question of cache locality 🤷🏼.
As for the diff size, it's pretty reasonable IMHO, I just took the opportunity to remove a left-over comment when `curr_cache` was still a thing and remove redundant `()` for the return tuple.